Break out fetches.
