+++ server/middleware/api.js
@@ -2,6 +2,8 @@
 import http from 'http';
 
 import mockRoomData from '../mocks/mock-data';
+import fetchRoomReservations from './fetch-room-reservation';
+import fetchRoomTemperature from './fetch-room-temperature';
 import setAlertByReservationStatus from './helpers/set-reservation-alert';
 import {
   MOCK_ROOM_RESERVATIONS,
@@ -10,43 +12,6 @@ import {
   EMIT_ROOM_STATUSES_UPDATE,
   EMIT_ROOM_TEMPERATURE_UPDATE
 } from '../ducks/rooms';
-
-const fetchRoomReservations = (next, action) => {
-  const { room, accessories } = action;
-  const source = `${FETCH_ROOM_RESERVATIONS}${room.outlookAccount}`;
-
-  // Retrieve outlook room reservation statuses
-  http.get(source, (response) => {
-    response.on('data', (data) => {
-      const reservations = JSON.parse(data.toString('utf8'));
-      const roomWithAlert = setAlertByReservationStatus(room, reservations);
-
-      next({
-        type: EMIT_ROOM_STATUSES_UPDATE,
-        room: roomWithAlert,
-        accessories
-      });
-    });
-  }).on('error', (error) => {
-    const errorMessage = `Failed to fetch room reservations
-                          for ${room.outlookAccount}. \n
-                          ${error}`;
-
-    console.log(errorMessage);
-  });
-};
-
-const fetchRoomTemperature = (room, next, action) => {
-  const { thermo } = action.accessories;
-
-  thermo.on('data', () => {
-    next({
-      type: EMIT_ROOM_TEMPERATURE_UPDATE,
-      room,
-      temperature: thermo.F
-    });
-  });
-};
 
 export default () => (next) => (action) => {
   const { room, accessories } = action;
